Refactor client code
