+++ client/src/main.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import * as ReactDOM from "react-dom";
 
-import App from "./App";
+import App from "./components/App";
 
 declare var __NAME__: string;
 declare var __VERSION__: string;
